% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthexplorer_download.R
\name{earthexplorer_download}
\alias{earthexplorer_download}
\title{earthexplorer_download}
\usage{
earthexplorer_download(usgs_eros_username, usgs_eros_password,
  output_folder = getwd(), earthexplorer_search_results = NULL,
  quicklooks_only = F, ordernum = NULL, overwrite = F, verbose = F)
}
\arguments{
\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{output_folder}{Character. Folder to save results to.}

\item{earthexplorer_search_results}{List. Output from earthexplorer_search or espa_inventory_search.  Only used to download quicklooks.}

\item{quicklooks_only}{Logical. If earthexplorer_search_results isn't NULL, will download the quicklooks of the search results.}

\item{ordernum}{Character. An order number. See https://espa.cr.usgs.gov/ordering/order-status/ for the IDs.}

\item{overwrite}{Logical. Overwrite existing results?  Default is FALSE}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
NULL and the files downloaded into the output_folder.
}
\description{
Downloads from Earth Explorer.
}
\details{
This is esssentially the command line version of the bulk download application.
Typically you submit a search using earthexplorer_search or just the Earth Explorer GUI, wait
a bit for it to process, and when you get your notificaiton the order is complete, you
submit the ordernum to this function for batch downloading.
}
\examples{
\dontrun{ 
# Get a list of your orders, and choose the first order result:
my_orders <- espa_ordering_listorders(usgs_eros_username="myusername",
	usgs_eros_password="mypassword",ordernums_only=T)
my_latest_order <- my_orders[1]
earthexplorer_download(usgs_eros_username="myusername",usgs_eros_password="mypassword",
	output_folder=getwd()
	ordernum=my_latest_order)
}
}
\references{
\url{https://earthexplorer.usgs.gov}
}
\seealso{
\code{\link{espa_inventory_search},\link{espa_ordering_order}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net})
}
