% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthexplorer_search.R
\name{earthexplorer_search}
\alias{earthexplorer_search}
\title{earthexplorer_search}
\usage{
earthexplorer_search(usgs_eros_username, usgs_eros_password, datasetName,
  lowerLeft = NULL, upperRight = NULL, startDate = "1920-01-07",
  endDate = as.character(Sys.Date()), months = "",
  includeUnknownCloudCover = T, minCloudCover = 0, maxCloudCover = 100,
  additionalCriteria = "", sp = NULL, place_order = F, products = "sr",
  format = "gtiff", verbose = F)
}
\arguments{
\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{datasetName}{Character. Used as a filter with wildcards inserted at the beginning and the end of the supplied value.  See https://mapbox.github.io/usgs/reference/catalog/hdds.html for a list of valid datasetNames.  See Details for additional info.}

\item{lowerLeft}{List. When used in conjunction with upperRight, creates a bounding box to search spatially. Should be a list of form list(latitude=XX,longitude=XX)}

\item{upperRight}{List. When used in conjunction with lowerLeft, creates a bounding box to search spatially. Should be a list of form list(latitude=XX,longitude=XX)}

\item{startDate}{Character. Used to search datasets temporally for possible dataset coverage. ISO 8601 Formatted Date. Time portion is ignored. Default is "1920-01-01".}

\item{endDate}{Character. Used to search datasets temporally for possible dataset coverage. ISO 8601 Formatted Date. Time portion is ignored.Default is today's date.}

\item{months}{Numeric. Used to limit results to specific months.}

\item{includeUnknownCloudCover}{Logical. Used to determine if scenes with unknown cloud cover values should be included in the results. Default is TRUE.}

\item{minCloudCover}{Numeric. Used to limit results by minimum cloud cover (for supported datasets). Range 0 to 100. Default is 0.}

\item{maxCloudCover}{Numeric. Used to limit results by maximum cloud cover (for supported datasets). Range 0 to 100. Default is 100.}

\item{additionalCriteria}{List. Used to filter results based on dataset specific metadata fields. Use datasetFields request to determine available fields and options.}

\item{sp}{SpatialPolygons. Used to refine the search results based on a polygon rather than a bounding box.}

\item{place_order}{Logical. If T, the search will be ordered.  Usually you want to search and refine first before ordering. Default is F.}

\item{products}{Character. A vector of products to be ordered.  Default is "sr" (surface reflectance).}

\item{format}{Character. The output file format for an order.  Default is "gtiff".}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of search results.
}
\description{
Searches and orders from Earth Explorer.
}
\details{
This is meant to be a fully functional command line interface to searching and (optionally) ordering products from Earth Explorer.

We have added some custom datasetNames to make searching a bit easier.  If datasetName="LANDSAT_4578", the search
will include Landsat 4 TM, Landsat 5 TM, Landset 7 ETM+, and Landsat 8 OLI.  This is a good if you want to order a long
time series of surface reflectance products.  Set products="sr" in that case.
}
\examples{
\dontrun{ 
# Search only:
earthexplorer_search(
	usgs_eros_username="myusername",usgs_eros_password="mypassword",
	datasetName="GLS2005",
		"lowerLeft"=list(latitude=75,longitude=-135),
		"upperRight"=list(latitude=90,longitude=-120),
		startDate="2006-01-01",endDate="2007-12-01",
	includeUnknownCloudCover=T,minCloudCover=0,maxCloudCover=100,
	place_order = F,
	verbose=T)

# Search and order:
earthexplorer_search(
	usgs_eros_username="myusername",usgs_eros_password="mypassword",
	datasetName="GLS2005",
		"lowerLeft"=list(latitude=75,longitude=-135),
		"upperRight"=list(latitude=90,longitude=-120),
		startDate="2006-01-01",endDate="2007-12-01",
	includeUnknownCloudCover=T,minCloudCover=0,maxCloudCover=100,
	place_order = T,
		products="sr",
		format="gtiff",
	verbose=T)
}
}
\references{
\url{https://earthexplorer.usgs.gov}
}
\seealso{
\code{\link{espa_inventory_search},\link{espa_ordering_order}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
