% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_datasetfields.R
\name{espa_inventory_datasetfields}
\alias{espa_inventory_datasetfields}
\title{espa_inventory_datasetfields}
\usage{
espa_inventory_datasetfields(datasetName, node = "EE", apiKey, verbose = F)
}
\arguments{
\item{datasetName}{Character.  Identifies the dataset.  See https://mapbox.github.io/usgs/reference/catalog/hdds.html for a list of valid datasetNames.}

\item{node}{Character. Determines the dataset catalog to use. Default="EE", probably never needs to be changed.}

\item{apiKey}{Character. Users API Key/Authentication Token, obtained from espa_inventory_login request.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of metadata filter fields for the specified dataset.
}
\description{
R wrapper for USGS/EROS Inventory Service Dataset Fields API.
}
\details{
This request is used to return the metadata filter fields for the specified dataset. 
These values can be used as additional criteria when submitting search and hit queries.
}
\examples{
\dontrun{ 
# Get API KEY (required EROS username/password): 
apiKey=espa_inventory_login("myusername","mypassword")
espa_inventory_datasetfields(datasetName="LANDSAT_8",apiKey=apiKey)
}
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api#datasetfields}
}
\seealso{
\code{\link{espa_inventory_login},\link{espa_inventory_get_api}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
