% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_datasets.R
\name{espa_inventory_datasets}
\alias{espa_inventory_datasets}
\title{espa_inventory_datasets}
\usage{
espa_inventory_datasets(datasetName = NULL, spatialFilter = NULL,
  temporalFilter = NULL, lowerLeft = NULL, upperRight = NULL,
  startDate = "1920-01-07", endDate = Sys.Date(), apiKey, verbose = F)
}
\arguments{
\item{datasetName}{Character. Used as a filter with wildcards inserted at the beginning and the end of the supplied value.  See https://mapbox.github.io/usgs/reference/catalog/hdds.html for a list of valid datasetNames.}

\item{lowerLeft}{List. When used in conjunction with upperRight, creates a bounding box to search spatially. Should be a list of form list(latitude=XX,longitude=XX)}

\item{upperRight}{List. When used in conjunction with lowerLeft, creates a bounding box to search spatially. Should be a list of form list(latitude=XX,longitude=XX)}

\item{startDate}{Character. Used to search datasets temporally for possible dataset coverage. ISO 8601 Formatted Date. Time portion is ignored.}

\item{endDate}{Character. Used to search datasets temporally for possible dataset coverage. ISO 8601 Formatted Date. Time portion is ignored.}

\item{apiKey}{Character. Users API Key/Authentication Token, obtained from espa_inventory_login request.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of metadata filter fields for datasets available for searching.
}
\description{
R wrapper for USGS/EROS Inventory Service Dataset Search API.
}
\details{
This method is used to find datasets available for searching. 
By passing no parameters except node, all available datasets are returned. 
Additional parameters such as temporal range and spatial bounding box can 
be used to find datasets that provide more specific data. The dataset name 
parameter can be used to limit the results based on matching the supplied 
value against the public dataset name with assumed wildcards at the beginning 
and end.
}
\examples{
\dontrun{ 
# Get API KEY (required EROS username/password): 
apiKey=espa_inventory_login("myusername","mypassword")
espa_inventory_datasets(datasetName="LANDSAT_8",lowerLeft=list(latitude=44.60847,"longitude"=-99.69639),
   upperRight=list(latitude=44.60847,"longitude"=-99.69639),
   startDate="2014-10-01",endDate="2014-10-01",
   apiKey=apiKey,
   node="EE") 
}
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api#datasets}
}
\seealso{
\code{\link{espa_inventory_login},\link{espa_inventory_get_api}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
