% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_get_api.R
\name{espa_inventory_get_api}
\alias{espa_inventory_get_api}
\title{espa_inventory_get_api}
\usage{
espa_inventory_get_api(request_code, json_request_content, auto_unbox = T,
  verbose = F)
}
\arguments{
\item{request_code}{Character. One of the requests listed at https://earthexplorer.usgs.gov/inventory/documentation/json-api}

\item{json_request_content}{List. Parameters to pass to the request system.}

\item{auto_unbox}{Logical auto_unbox the json_request_content?  Default = T.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List generated from the request.
}
\description{
R wrapper for USGS/EROS Inventory Service GET API.
}
\details{
A function to convert requests to the USGS/EROS Inventory Service Request URL format (GET).
In general, you shouldn't need to call this directly unless you want finer command-line control over requests.
}
\examples{
\dontrun{ 
# Run the simple API status command:
espa_inventory_get_api(request_code="status",json_request_content="",verbose=F)
}
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api}
}
\seealso{
\code{\link{espa_inventory_post_api}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
