% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_grid2ll.R
\name{espa_inventory_grid2ll}
\alias{espa_inventory_grid2ll}
\title{espa_inventory_datasets}
\usage{
espa_inventory_grid2ll(gridType, responseShape, path, row, verbose = F)
}
\arguments{
\item{gridType}{Character. Identifies the grid system    Accepted values: "WRS1" and "WRS2".}

\item{responseShape}{Character. Determines if the response should be a center point or outer polygon    Accepted values: "point" and "polygon"}

\item{path}{Numeric. WRS Path.}

\item{row}{Numeric. WRS Row.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of metadata on the conversion of the grid systems to lat/lng center points or polygons.
}
\description{
R wrapper for USGS/EROS Inventory Service Dataset Grid to Lat/Lng API.
}
\details{
This method is used to convert the following grid systems to lat/lng 
center points or polygons: WRS-1, WRS-2. To account for multiple grid 
systems there are required fields for all grids (see "Required Request 
Parameters") as well as grid-specific parameters.
}
\examples{
\dontrun{ 
espa_inventory_grid2ll(gridType="WRS1",responseShape="polygon",path=1,row=1)
}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api#grid2ll}
}
\seealso{
\code{\link{espa_inventory_get_api}}
}

