% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_hits.R
\name{espa_inventory_hits}
\alias{espa_inventory_hits}
\title{espa_inventory_hits}
\usage{
espa_inventory_hits(datasetName, lowerLeft = "", upperRight = "",
  startDate = "1920-01-07", endDate = Sys.Date(), months = "",
  includeUnknownCloudCover = T, minCloudCover = 0, maxCloudCover = 100,
  additionalCriteria = "", apiKey, node = "EE", verbose = F)
}
\arguments{
\item{datasetName}{Character. Used as a filter with wildcards inserted at the beginning and the end of the supplied value.  See https://mapbox.github.io/usgs/reference/catalog/hdds.html for a list of valid datasetNames.}

\item{lowerLeft}{List. When used in conjunction with upperRight, creates a bounding box to search spatially. Should be a list of form list(latitude=XX,longitude=XX)}

\item{upperRight}{List. When used in conjunction with lowerLeft, creates a bounding box to search spatially. Should be a list of form list(latitude=XX,longitude=XX)}

\item{startDate}{Character. Used to search datasets temporally for possible dataset coverage. ISO 8601 Formatted Date. Time portion is ignored. Default is "1920-01-01".}

\item{endDate}{Character. Used to search datasets temporally for possible dataset coverage. ISO 8601 Formatted Date. Time portion is ignored.Default is today's date.}

\item{months}{Numeric. Used to limit results to specific months.}

\item{includeUnknownCloudCover}{Logical. Used to determine if scenes with unknown cloud cover values should be included in the results. Default is TRUE.}

\item{minCloudCover}{Numeric. Used to limit results by minimum cloud cover (for supported datasets). Range 0 to 100. Default is 0.}

\item{maxCloudCover}{Numeric. Used to limit results by maximum cloud cover (for supported datasets). Range 0 to 100. Default is 100.}

\item{additionalCriteria}{List. Used to filter results based on dataset specific metadata fields. Use datasetFields request to determine available fields and options.}

\item{apiKey}{Character. Users API Key/Authentication Token, obtained from espa_inventory_login request.}

\item{node}{Character. Determines the dataset catalog to use. Default="EE", probably never needs to be changed.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
A list contains the results of the search.
}
\description{
R wrapper for USGS/EROS Inventory Service Scene Search Hits API.
}
\details{
This method is used in determining the number of hits a search returns. 
Because a hits request requires a search, this request takes the same 
parameters as the search request, with exception to the non-search-field 
parameters; maxResults, startingNumber, and sortOrder.
}
\examples{
\dontrun{ 
# Get API KEY (required EROS username/password): 
apiKey=espa_inventory_login("myusername","mypassword")
# Standard search:
espa_inventory_hits(datasetName="GLS2005",
		"lowerLeft"=list(latitude=75,longitude=-135),
		"upperRight"=list(latitude=90,longitude=-120),
		startDate="2006-01-01",endDate="2007-12-01",
		node="EE",
		apiKey=apiKey)
}
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api#hits}
}
\seealso{
\code{\link{espa_inventory_login},\link{espa_inventory_get_api}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
