% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_login.R
\name{espa_inventory_login}
\alias{espa_inventory_login}
\title{espa_inventory_login}
\usage{
espa_inventory_login(usgs_eros_username, usgs_eros_password, verbose = F)
}
\arguments{
\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
Character of the API KEY needed by the other inventory APIs.
}
\description{
R wrapper for USGS/EROS Inventory Service Login API.
}
\details{
Upon a successful login, an API key will be returned. 
This key will be active for one hour and should be 
destroyed upon final use of the service by calling the 
logout method. Users must have "Machine to Machine" 
access, which is established in the user's profile.
}
\examples{
\dontrun{ 
# Get API KEY (required EROS username/password): 
apiKey=espa_inventory_login("myusername","mypassword")
}
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api#login}
}
\seealso{
\code{\link{espa_inventory_logout}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
