% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_metadata.R
\name{espa_inventory_metadata}
\alias{espa_inventory_metadata}
\title{espa_inventory_metadata}
\usage{
espa_inventory_metadata(datasetName, entityIds, apiKey, node = "EE",
  verbose = F)
}
\arguments{
\item{datasetName}{Character. Identifies the dataset    Use the datasetName from datasets response.  See https://mapbox.github.io/usgs/reference/catalog/hdds.html for a list of valid datasetNames.}

\item{entityIds}{Character. Identifies multiple scenes to remove.}

\item{apiKey}{Character. Users API Key/Authentication Token, obtained from espa_inventory_login request.}

\item{node}{Character. Determines the dataset catalog to use. Default="EE", probably never needs to be changed.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of metadata.
}
\description{
R wrapper for USGS/EROS Inventory Service Scene Metadata API.
}
\details{
The use of the metadata request is intended for those who have 
acquired scene IDs from a different source. It will return the 
same metadata that is available via the search request.
}
\examples{
\dontrun{ 
# Get API KEY (required EROS username/password): 
apiKey=espa_inventory_login("myusername","mypassword")
espa_inventory_metadata(datasetName="LANDSAT_8",entityIds=c("LC80130292014100LGN00"),apiKey=apiKey)
}
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api#metadata}
}
\seealso{
\code{\link{espa_inventory_login},\link{espa_inventory_get_api}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
