% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_post_api.R
\name{espa_inventory_post_api}
\alias{espa_inventory_post_api}
\title{espa_inventory_post_api}
\usage{
espa_inventory_post_api(request_code, json_request_content, verbose = F)
}
\arguments{
\item{request_code}{Character. One of the requests listed at https://earthexplorer.usgs.gov/inventory/documentation/json-api}

\item{json_request_content}{List. Parameters to pass to the request system.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List generated from the request.
}
\description{
R wrapper for USGS/EROS Inventory Service POST API.
}
\details{
A function to convert requests to the USGS/EROS Inventory Service Request URL format (POST).
In general, you shouldn't need to call this directly unless you want finer command-line control over requests.
}
\examples{
\dontrun{ 
# Run the API retrieval command:
	login_parameters <- list(username="myusername",password="mypassword")
	espa_inventory_post_api(request_code="login",
    json_request_content=login_parameters,verbose=verbose)
}
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api}
}
\seealso{
\code{\link{espa_inventory_login},\link{espa_inventory_get_api}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
