% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_spatialFilter.R
\name{espa_inventory_spatialFilter}
\alias{espa_inventory_spatialFilter}
\title{espa_inventory_spatialFilter}
\usage{
espa_inventory_spatialFilter(lowerLeft, upperRight, filterType = "mbr")
}
\arguments{
\item{lowerLeft}{List. When used in conjunction with upperRight, creates a bounding box to search spatially. Should be a list of form list(latitude=XX,longitude=XX)}

\item{upperRight}{List. When used in conjunction with lowerLeft, creates a bounding box to search spatially. Should be a list of form list(latitude=XX,longitude=XX)}

\item{filterType}{TBD.}
}
\description{
espa_inventory_spatialFilter
}
\details{
TBD
}
\examples{
\dontrun{ 
### TBD
}

}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net})
}
