% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_status.R
\name{espa_inventory_status}
\alias{espa_inventory_status}
\title{espa_inventory_status}
\usage{
espa_inventory_status(verbose = F)
}
\arguments{
\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of status info of the API.
}
\description{
R wrapper for USGS/EROS Inventory Service Status API.
}
\details{
This method is used to get the status of the API.
}
\examples{
\dontrun{ 
espa_inventory_status()
}
}
\references{
\url{https://earthexplorer.usgs.gov/inventory/documentation/json-api#status}
}
\seealso{
\code{\link{espa_inventory_get_api}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
