% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_inventory_temporalFilter.R
\name{espa_inventory_temporalFilter}
\alias{espa_inventory_temporalFilter}
\title{espa_inventory_temporalFilter}
\usage{
espa_inventory_temporalFilter(startDate, endDate, dateField = "search_date")
}
\arguments{
\item{startDate}{TBD}

\item{endDate}{TBD}

\item{dateField}{TBD}
}
\description{
espa_inventory_temporalFilter
}
\details{
TBD
}
\examples{
\dontrun{ 
### TBD
}

}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net})
}
