% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_ordering_apioperations.R
\name{espa_ordering_apioperations}
\alias{espa_ordering_apioperations}
\title{espa_ordering_apioperations}
\usage{
espa_ordering_apioperations(usgs_eros_username, usgs_eros_password,
  verbose = F)
}
\arguments{
\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of available api operations.
}
\description{
R wrapper for ESPA Ordering System API Operations API.
}
\details{
Lists all available api operations.
}
\examples{
\dontrun{ 
espa_ordering_apioperations(usgs_eros_username="myusername",usgs_eros_password="mypassword")
}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
\references{
\url{https://github.com/USGS-EROS/espa-api#apiOps}
}

