% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_ordering_itemstatus.R
\name{espa_ordering_itemstatus}
\alias{espa_ordering_itemstatus}
\title{espa_ordering_itemstatus}
\usage{
espa_ordering_itemstatus(ordernum = NULL, itemnum = NULL,
  usgs_eros_username, usgs_eros_password, verbose = F)
}
\arguments{
\item{ordernum}{Character. An order number. See https://espa.cr.usgs.gov/ordering/order-status/ for the IDs.}

\item{itemnum}{Character. An item number.}

\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of status and details of items in an order.
}
\description{
R wrapper for ESPA Ordering System Status and Details API.
}
\details{
Retrieve status and details for all or a particular product in an order
}
\examples{
\dontrun{ 
# Get status for all items in a given order:
espa_ordering_itemstatus(ordernum="my@email.com-01122017-154454-218",usgs_eros_username="myusername",usgs_eros_password="mypassword")

# Get status for a single item in a given order:
espa_ordering_itemstatus(ordernum="my@email.com-01122017-154454-218",itemnum="LC80410312015219LGN00",usgs_eros_username="myusername",usgs_eros_password="mypassword")
}
}
\references{
\url{https://github.com/USGS-EROS/espa-api#apiProdStats}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
