% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_ordering_listorders.R
\name{espa_ordering_listorders}
\alias{espa_ordering_listorders}
\title{espa_ordering_listorders}
\usage{
espa_ordering_listorders(email = NULL, usgs_eros_username, usgs_eros_password,
  ordernums_only = F, verbose = F)
}
\arguments{
\item{email}{Character. email address of a user.  Only neccessary if you want to list orders from a different user.}

\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{ordernums_only}{Logical. Returns the orders in a nearly formatted character vector. Default=F.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
(ordernums_only=F) list or (ordernums_only=T) character vector of orders for a given user.
}
\description{
R wrapper for ESPA Ordering System List Orders API.
}
\details{
Lists orders for the authenticated user or a supplied email.
}
\examples{
\dontrun{ 
# Get status for the authenticated user:
espa_ordering_listorders(usgs_eros_username="myusername",usgs_eros_password="mypassword")

# Get status for a different user based on their email:
espa_ordering_listorders(email="my@email.com",usgs_eros_username="myusername",usgs_eros_password="mypassword")
}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
\references{
\url{https://github.com/USGS-EROS/espa-api#apiOrdersEmail}
}

