% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_ordering_order.R
\name{espa_ordering_order}
\alias{espa_ordering_order}
\title{espa_ordering_order}
\usage{
espa_ordering_order(ordernum = NULL, orderrequest = NULL,
  usgs_eros_username, usgs_eros_password, verbose = F)
}
\arguments{
\item{ordernum}{Character. An order number. See https://espa.cr.usgs.gov/ordering/order-status/ for the IDs.}

\item{orderrequest}{List. A properly formatted order request.  See the examples for details.}

\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
If ordernum is supplied, a list of details of items in an order. If orderrequest is supplied, an ordernum.
}
\description{
R wrapper for ESPA Ordering System Submit Order or Retrieve Details of an Order API.
}
\details{
Submit or retrieves details of an order with ESPA.
}
\examples{
\dontrun{ 
orderrequest=list(
	olitirs8=list(
			inputs="LC80270292015233LGN00",products="sr"),
	format=unbox("gtiff")) 
myorder <- espa_ordering_order(orderrequest=orderrequest,
		usgs_eros_username="myusername",usgs_eros_password="mypassword")
}
}
\references{
\url{https://github.com/USGS-EROS/espa-api#apiSubmitOrder}
}
\seealso{
\code{\link{earthexplorer_search}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
