% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_ordering_orderingschema.R
\name{espa_ordering_orderingschema}
\alias{espa_ordering_orderingschema}
\title{espa_ordering_orderingschema}
\usage{
espa_ordering_orderingschema(usgs_eros_username, usgs_eros_password,
  verbose = F)
}
\arguments{
\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List of order schema.
}
\description{
R wrapper for ESPA Ordering System Retrieve Order Schema API.
}
\details{
Retrieves order schema definition
}
\examples{
\dontrun{ 
espa_ordering_orderingschema(usgs_eros_username="myusername",usgs_eros_password="mypassword")
}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
\references{
\url{https://github.com/USGS-EROS/espa-api#apiOrderSchema}
}

