% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espa_ordering_post_api.R
\name{espa_ordering_post_api}
\alias{espa_ordering_post_api}
\title{espa_ordering_post_api}
\usage{
espa_ordering_post_api(request_code, json_request_content = NULL,
  usgs_eros_username, usgs_eros_password, auto_unbox = T, verbose = F)
}
\arguments{
\item{request_code}{Character. One of the requests listed at https://earthexplorer.usgs.gov/inventory/documentation/json-api}

\item{json_request_content}{List. Parameters to pass to the request system.}

\item{usgs_eros_username}{Character. Your USGS registration username. You can register for an account at https://espa.cr.usgs.gov}

\item{usgs_eros_password}{Character. Your USGS registration password. WARNING: USING THIS WITH R IS NOT PARTICULARLY SECURE.}

\item{auto_unbox}{Logical auto_unbox the json_request_content?  Default = T.}

\item{verbose}{Logical. Verbose execution?  Default=F.}
}
\value{
List generated from the request.
}
\description{
R wrapper for USGS/EROS Ordering System POST API.
}
\details{
A function to convert requests to the USGS/EROS Ordering System URL format (POST).
In general, you shouldn't need to call this directly unless you want finer command-line control over requests.
}
\examples{
\dontrun{ 
# Run the Available Products command:
inputs <- c("LE70290302003123EDC00","MOD09A1.A2000073.h12v11.005.2008238080250.hdf","bad_scene_id")
json_request_content <- list(inputs=inputs)
espa_ordering_post_api(request_code="api/v0/available-products",json_request_content=inputs,
  usgs_eros_username="myusername",usgs_eros_password="mypassword")
}
}
\seealso{
\code{\link{espa_ordering_get_api}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net}) (wrapper) 
and USGS ESPA Developers (API and documentation)
}
