% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_to_spatialPolygons.R
\name{search_to_spatialPolygons}
\alias{search_to_spatialPolygons}
\title{search_to_spatialPolygons}
\usage{
search_to_spatialPolygons(espa_inventory_search_results)
}
\arguments{
\item{espa_inventory_search_results}{List. The outputs of espa_inventory_search}
}
\value{
Character of the API KEY needed by the other inventory APIs.
}
\description{
Converts espa inventory search results to a SpatialPolygonsDataFrame
}
\examples{
\dontrun{ 
# Get API KEY (required EROS username/password): 
apiKey=espa_inventory_login("myusername","mypassword")

# Standard search:
search_results <- espa_inventory_search(datasetName="GLS2005",
		"lowerLeft"=list(latitude=75,longitude=-135),
		"upperRight"=list(latitude=90,longitude=-120),
		startDate="2006-01-01",endDate="2007-12-01",
		node="EE",
		maxResults=3,startingNumber=1,sortOrder="ASC",
		apiKey=apiKey)

plot(search_to_spatialPolygons(espa_inventory_search_results=search_results))
}
}
\seealso{
\code{\link{espa_inventory_search}}
}
\author{
Jonathan A. Greenberg (\email{espa-tools@estarcion.net})
}
